Change osTeleportAgent parameters from long to int. That numerical range is not even supported by the underlying type, so there is no need to ask for a type the script can not even supply.

+++ OpenSim/Region/ScriptEngine/Shared/Api/Runtime/OSSL_Stub.cs
@@ -201,9 +201,9 @@ namespace OpenSim.Region.ScriptEngine.Shared.ScriptBase
             m_OSSL_Functions.osTeleportAgent(agent, regionName, position, lookat);
         }
 
-        public void osTeleportAgent(string agent, long regionX, long regionY, vector position, vector lookat)
+        public void osTeleportAgent(string agent, int regionX, int regionY, vector position, vector lookat)
         {
-            m_OSSL_Functions.osTeleportAgent(agent, (uint) regionX, (uint) regionY, position, lookat);
+            m_OSSL_Functions.osTeleportAgent(agent, regionX, regionY, position, lookat);
         }
 
         public void osTeleportAgent(string agent, vector position, vector lookat)
